marites-worker/requirements.txt
pyTigerGraph
python-dotenv
